# Increase HTTP/2 default flow control window.