--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -23,6 +23,7 @@
   We expect this to be uncommon, but please make yourself heard if that's not the case.
   ([#7286](https://github.com/mitmproxy/mitmproxy/pull/7286), @lukant)
 - Increase HTTP/2 default flow control window size.
+  ([#7317](https://github.com/mitmproxy/mitmproxy/pull/7317), @sujaldev)
 
 ## 02 October 2024: mitmproxy 11.0.0