--- a/mitmproxy/proxy/layers/http/_http_h2.py
+++ b/mitmproxy/proxy/layers/http/_http_h2.py
@@ -55,6 +55,11 @@ class BufferedH2Connection(h2.connection.H2Connection):
         self.stream_buffers = collections.defaultdict(collections.deque)
         self.stream_trailers = {}
 
+    def initiate_connection(self):
+        super().initiate_connection()
+        # The connection flow-control window can only be changed using WINDOW_UPDATE frames.
+        self.increment_flow_control_window(2 ** 24)
+
     def send_data(
         self,
         stream_id: int,