--- a/mitmproxy/proxy/layers/http/_http_h2.py
+++ b/mitmproxy/proxy/layers/http/_http_h2.py
@@ -58,7 +58,7 @@ class BufferedH2Connection(h2.connection.H2Connection):
     def initiate_connection(self):
         super().initiate_connection()
         # The connection flow-control window can only be changed using WINDOW_UPDATE frames.
-        self.increment_flow_control_window(2 ** 24)
+        self.increment_flow_control_window(2**24)
 
     def send_data(
         self,