--- a/mitmproxy/proxy/layers/http/_http_h2.py
+++ b/mitmproxy/proxy/layers/http/_http_h2.py
@@ -49,7 +49,7 @@ class BufferedH2Connection(h2.connection.H2Connection):
 
     def __init__(self, config: h2.config.H2Configuration):
         super().__init__(config)
-        self.local_settings.initial_window_size = 2**24
+        self.local_settings.initial_window_size = 2**31-1
         self.local_settings.max_frame_size = 2**17
         self.max_inbound_frame_size = 2**17
         # hyper-h2 pitfall: we need to acknowledge here, otherwise its sends out the old settings.