--- a/mitmproxy/proxy/layers/http/_http_h2.py
+++ b/mitmproxy/proxy/layers/http/_http_h2.py
@@ -61,7 +61,7 @@ class BufferedH2Connection(h2.connection.H2Connection):
         super().initiate_connection()
         # We increase the flow-control window for new streams with a setting,
         # but we need to increase the overall connection flow-control window as well.
-        self.increment_flow_control_window(2**24)
+        self.increment_flow_control_window(2**31 - 1)
 
     def send_data(
         self,