#### Description

Increases the default flow control window for http2 connections, fixes #7271.

#### Checklist

 - [x] I have updated tests where applicable.
 - [x] I have added an entry to the CHANGELOG.