@mhils 2024-11-14
Thanks! Can you report preliminary performance results w.r.t. #7271?