@sujaldev 2024-11-14
Downloading [downloads.mitmproxy.org/11.0.0/mitmproxy-11.0.0-linux-x86_64.tar.gz](https://downloads.mitmproxy.org/11.0.0/mitmproxy-11.0.0-linux-x86_64.tar.gz) with curl on fedora 41:
|                       Command                        | Time (seconds) |
|:----------------------------------------------------:|:--------------:|
|                  without mitmproxy                   |       11       |
|                      `mitmdump`                      |       14       |
|        `mitmdump --set stream_large_bodies=0`        |       13       |
|               `mitmdump --mode socks5`               |       14       |
| `mitmdump --mode socks5 --set stream_large_bodies=0` |       11       |