@mhils 2024-11-14
Thanks. This looks really impressive, but for some weird reason I get the opposite effect with this PR applied. Could you try and see what kind of data you get with https://github.com/mitmproxy/proxybench?
Here's my machine on git main:
```
proxy/bench/EncryptedHttp1 req=100mb resp=100mb, w=1 reqs=1 proxy=/
                        time:   [58.805 ms 61.606 ms 65.476 ms]
                        thrpt:  [15.273  elem/s 16.232  elem/s 17.005  elem/s]

proxy/bench/EncryptedHttp1 req=100mb resp=100mb, w=1 reqs=1 proxy=8080
                        time:   [982.66 ms 1.0673 s 1.1629 s]
                        thrpt:  [0.8599  elem/s 0.9370  elem/s 1.0176  elem/s]

proxy/bench/EncryptedHttp2 req=100mb resp=100mb, w=1 reqs=1 proxy=/
                        time:   [66.921 ms 70.992 ms 75.800 ms]
                        thrpt:  [13.193  elem/s 14.086  elem/s 14.943  elem/s]

proxy/bench/EncryptedHttp2 req=100mb resp=100mb, w=1 reqs=1 proxy=8080
                        time:   [2.9815 s 3.1839 s 3.3810 s]
                        thrpt:  [0.2958  elem/s 0.3141  elem/s 0.3354  elem/s]
```
And with your PR applied:
```
proxy/bench/EncryptedHttp1 req=100mb resp=100mb, w=1 reqs=1 proxy=/
                        time:   [56.542 ms 57.835 ms 59.096 ms]
                        thrpt:  [16.921  elem/s 17.291  elem/s 17.686  elem/s]

proxy/bench/EncryptedHttp1 req=100mb resp=100mb, w=1 reqs=1 proxy=8080
                        time:   [964.63 ms 1.0194 s 1.0769 s]
                        thrpt:  [0.9286  elem/s 0.9809  elem/s 1.0367  elem/s]

proxy/bench/EncryptedHttp2 req=100mb resp=100mb, w=1 reqs=1 proxy=/
                        time:   [53.895 ms 56.618 ms 59.534 ms]
                        thrpt:  [16.797  elem/s 17.662  elem/s 18.555  elem/s]

proxy/bench/EncryptedHttp2 req=100mb resp=100mb, w=1 reqs=1 proxy=8080
[stuck for at least a minute reading the request, never manages to complete a single one?]
```

(Note that it's also a possibility that the benchmarking tool is broken)