@sujaldev 2024-11-15
Here's what I get on my machine ([full log](https://files.sujal.dev/log.txt)):

#### `main` branch

```
proxy/bench/EncryptedHttp1 req=100mb resp=100mb, w=1 reqs=1 proxy=/                                                                          
                        time:   [161.74 ms 175.68 ms 188.68 ms]
                        thrpt:  [5.2999  elem/s 5.6921  elem/s 6.1827  elem/s]

proxy/bench/EncryptedHttp1 req=100mb resp=100mb, w=1 reqs=1 proxy=8080                                                                          
                        time:   [1.0207 s 1.0635 s 1.1143 s]
                        thrpt:  [0.8974  elem/s 0.9403  elem/s 0.9797  elem/s]

proxy/bench/EncryptedHttp2 req=100mb resp=100mb, w=1 reqs=1 proxy=/                                                                          
                        time:   [112.21 ms 117.76 ms 123.62 ms]
                        thrpt:  [8.0893  elem/s 8.4916  elem/s 8.9119  elem/s]

proxy/bench/EncryptedHttp2 req=100mb resp=100mb, w=1 reqs=1 proxy=8080                                                                          
                        time:   [2.4506 s 7.8673 s 18.610 s]
                        thrpt:  [0.0537  elem/s 0.1271  elem/s 0.4081  elem/s]
```


#### this PR:

```
proxy/bench/EncryptedHttp1 req=100mb resp=100mb, w=1 reqs=1 proxy=/                                                                          
                        time:   [179.46 ms 190.51 ms 200.88 ms]
                        thrpt:  [4.9781  elem/s 5.2491  elem/s 5.5723  elem/s]

proxy/bench/EncryptedHttp1 req=100mb resp=100mb, w=1 reqs=1 proxy=8080                                                                          
                        time:   [1.0704 s 1.1508 s 1.2428 s]
                        thrpt:  [0.8046  elem/s 0.8690  elem/s 0.9342  elem/s]

proxy/bench/EncryptedHttp2 req=100mb resp=100mb, w=1 reqs=1 proxy=/                                                                          
                        time:   [119.08 ms 123.10 ms 127.36 ms]
                        thrpt:  [7.8516  elem/s 8.1237  elem/s 8.3976  elem/s]

Benchmarking proxy/bench/EncryptedHttp2 req=100mb resp=100mb, w=1 reqs=1 proxy=8080: Warming up for 10.000 ms
```

After that it gets stuck, wireshark shows PING frames being sent which aren't destined towards the proxy (I changed mitmdump's listen host to a different loopback address to distinguish traffic).