@sujaldev 2024-11-16
I think the initial window size setting wasn't being updated in all scenarios with the previous approach, with the new changes proxybench measures a significant performance improvement:

```
proxy/bench/EncryptedHttp2 req=100mb resp=100mb, w=1 reqs=1 proxy=8080
                        time:   [1.9286 s 1.9840 s 2.0370 s]
                        thrpt:  [0.4909  elem/s 0.5040  elem/s 0.5185  elem/s]
                 change:
                        time:   [-96.043% -94.499% -90.582%] (p = 0.00 < 0.05)
                        thrpt:  [+961.84% +1717.7% +2427.0%]
                        Performance has improved.
```