@sujaldev 2024-11-16
> Did you take a look at `MAX_FRAME_SIZE` as well by any chance? The default sits at just 16kb, may be worth bumping that a bit as well? Maybe not 256xing it like the window size, but maybe to 128kb or so?

Increasing the `MAX_FRAME_SIZE` seems to cause a `FRAME_SIZE_ERROR`, looking into this.