@sujaldev 2024-11-18
Should we perhaps use `H2Connection.inbound_flow_control_window` instead of manually setting this value?