@mhils 2024-11-18
That's a good suggestion! Do you want to open a PR with just that? :)